==> notebooks/1.0-rs-feature_engineering.ipynb <==
import sys
sys.path.append('C:/Users/visha/Documents/GitHub/Club_Mahindra_Data_Hack/Club_Mahindra_Data_Hack')

# Import standard libraries
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline
color = sns.color_palette()

# Import local libraries
from src.data import make_dataset
from src.features import build_features

pd.options.display.max_columns = 100
import warnings
warnings.filterwarnings("ignore")

### Build Data to work on

# Load original data file
raw_data = pd.read_csv(r'../data/raw/train_5CLrC8b/train.csv')
raw_data.shape

raw_data.head()

# Take a slice (60K observations) of the original data to train the model
interim_data = make_dataset.slice_data(raw_data)

# Split the Data into training, validation and test set
train, val, test = make_dataset.split_train_val_test(interim_data)

# build date related features on all the three datasets
train_df = build_features.date_related_features(train)
val_df = build_features.date_related_features(val)
test_df = build_features.date_related_features(test)

### Build 1st Set of Features

# build features by aggregating values of different columns grouped by memberid
mini_train_df = build_features.aggregate_features_by_memberid(train_df)
mini_val_df = build_features.aggregate_features_by_memberid(val_df)
mini_test_df = build_features.aggregate_features_by_memberid(test_df)

# build features by taking cumulative values of different columns grouped by memberid
mini_train_df_1 = build_features.cumulative_features_by_memberid(train_df, mini_train_df)
mini_val_df_1 = build_features.cumulative_features_by_memberid(val_df, mini_val_df)
mini_test_df_1 = build_features.cumulative_features_by_memberid(test_df, mini_test_df)

# build features by taking time gap on corresponding date features
mini_train_df_2 = build_features.time_gap_shift_1_features(train_df, mini_train_df_1)
mini_val_df_2 = build_features.time_gap_shift_1_features(val_df, mini_val_df_1)
mini_test_df_2 = build_features.time_gap_shift_1_features(test_df, mini_test_df_1)

# build more features by taking time gap on corresponding date features
mini_train_df_3 = build_features.time_gap_shift_2_features(train_df, mini_train_df_2)
mini_val_df_3 = build_features.time_gap_shift_2_features(val_df, mini_val_df_2)
mini_test_df_3 = build_features.time_gap_shift_2_features(test_df, mini_test_df_2)

# build features by taking difference in values of columns for each visit
mini_train_df_4 = build_features.inter_visit_features(train_df, mini_train_df_3)
mini_val_df_4 = build_features.inter_visit_features(val_df, mini_val_df_3)
mini_test_df_4 = build_features.inter_visit_features(test_df, mini_test_df_3)

# # build features by pivoting data on memberid and some specific columns
# mini_train_df_5 = build_features.pivot_features(train_df, mini_train_df_4)
# mini_val_df_5 = build_features.pivot_features(val_df, mini_val_df_4)
# mini_test_df_5 = build_features.pivot_features(test_df, mini_test_df_4)

# Remove initial columns from mini_data
columns_to_remove = ['memberid','resort_id','state_code_residence','checkin_date','booking_date']
mini_train_df_5 = mini_train_df_4.drop(columns_to_remove, axis=1)
mini_val_df_5 = mini_val_df_4.drop(columns_to_remove, axis=1)
mini_test_df_5 = mini_test_df_4.drop(columns_to_remove, axis=1)

# Check shapes of each modified dataset
print(mini_train_df_5.shape)
print(mini_val_df_5.shape)
print(mini_test_df_5.shape)

# Save the 1st set of built features to project drive
mini_train_df_5.to_csv(r'../data/processed/built_features/train_feat_set_1.csv', index=False)
mini_val_df_5.to_csv(r'../data/processed/built_features/val_feat_set_1.csv', index=False)
mini_test_df_5.to_csv(r'../data/processed/built_features/test_feat_set_1.csv', index=False)

### Build More Features

# build ratio_features on all three datasets
train_df_1 = build_features.ratio_features(train_df)
val_df_1 = build_features.ratio_features(val_df)
test_df_1 = build_features.ratio_features(test_df)

# build concatenated features on all three datasets
train_df_2 = build_features.concatenated_features(train_df_1)
val_df_2 = build_features.concatenated_features(val_df_1)
test_df_2 = build_features.concatenated_features(test_df_1)

# Merge the 1st set of features with the modified datasets 
train_df_3 = pd.merge(train_df_2, mini_train_df_5, on='reservation_id', how='left')
val_df_3 = pd.merge(val_df_2, mini_val_df_5, on='reservation_id', how='left')
test_df_3 = pd.merge(test_df_2, mini_test_df_5, on='reservation_id', how='left')

# Save the 2nd set of built features to project drive
train_df_3.to_csv(r'../data/processed/prepared_data/train_prepared.csv', index=False)
val_df_3.to_csv(r'../data/processed/prepared_data/val_prepared.csv', index=False)
test_df_3.to_csv(r'../data/processed/prepared_data/test_prepared.csv', index=False)
