==> notebooks/2.0-rs-data_processing.ipynb <==
import sys
sys.path.append('C:/Users/visha/Documents/GitHub/Club_Mahindra_Data_Hack/Club_Mahindra_Data_Hack')

# Import standard libraries
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline
color = sns.color_palette()

# Import local libraries
from src.data_preprocessing import preprocess_data

pd.options.display.max_columns = 100
import warnings
warnings.filterwarnings("ignore")

### Load Data

train_prepared = pd.read_csv(r'../data/processed/prepared_data/train_prepared.csv')
val_prepared = pd.read_csv(r'../data/processed/prepared_data/val_prepared.csv')
test_prepared = pd.read_csv(r'../data/processed/prepared_data/test_prepared.csv')

### Pre-process Data

# separate features from targets
train_prepared_X, train_prepared_y = preprocess_data.separate_features_and_targets(train_prepared)
val_prepared_X, val_prepared_y = preprocess_data.separate_features_and_targets(val_prepared)
test_prepared_X, test_prepared_y = preprocess_data.separate_features_and_targets(test_prepared)

# separate numerical features from categorical features
train_num_feature_names, train_cat_feature_names = preprocess_data.separate_num_cat_features(train_prepared_X)
val_num_feature_names, val_cat_feature_names = preprocess_data.separate_num_cat_features(val_prepared_X)
test_num_feature_names, test_cat_feature_names = preprocess_data.separate_num_cat_features(test_prepared_X)

# transform features to make it ready to be fed to machine learning models
train_preprocessed = preprocess_data.transform_all_features(train_prepared_X, train_num_feature_names, train_cat_feature_names)
val_preprocessed = preprocess_data.transform_all_features(val_prepared_X, val_num_feature_names, val_cat_feature_names)
test_preprocessed = preprocess_data.transform_all_features(test_prepared_X, test_num_feature_names, test_cat_feature_names)

# Convert the arrays into pandas DataFrame
train_preprocessed = pd.DataFrame(train_preprocessed)
val_preprocessed = pd.DataFrame(val_preprocessed)
test_preprocessed = pd.DataFrame(test_preprocessed)

# Save the preprocessed features to project drive
train_preprocessed.to_csv(r'../data/processed/preprocessed_data/train_preprocessed.csv', index=False)
val_preprocessed.to_csv(r'../data/processed/preprocessed_data/val_preprocessed.csv', index=False)
test_preprocessed.to_csv(r'../data/processed/preprocessed_data/test_preprocessed.csv', index=False)

# Save target values to project drive
train_prepared_y.to_csv(r'../data/processed/target_values/train_target_values.csv', index=False)
val_prepared_y.to_csv(r'../data/processed/target_values/val_target_values.csv', index=False)
test_prepared_y.to_csv(r'../data/processed/target_values/test_target_values.csv', index=False)
