==> notebooks/3.0-rs-model_building.ipynb <==
import sys
sys.path.append('C:/Users/visha/Documents/GitHub/Club_Mahindra_Data_Hack/Club_Mahindra_Data_Hack')

# Import standard libraries
import numpy as np
import pandas as pd
import pickle
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline
color = sns.color_palette()

# Import local libraries
from src.models import train_models

pd.options.display.max_columns = 100
import warnings
warnings.filterwarnings("ignore")

### Load Data

train_preprocessed = pd.read_csv(r'../data/processed/preprocessed_data/train_preprocessed.csv')
val_preprocessed = pd.read_csv(r'../data/processed/preprocessed_data/val_preprocessed.csv')
test_preprocessed = pd.read_csv(r'../data/processed/preprocessed_data/test_preprocessed.csv')

train_target = pd.read_csv(r'../data/processed/target_values/train_target_values.csv', header=None)
val_target = pd.read_csv(r'../data/processed/target_values/val_target_values.csv', header=None)
test_target = pd.read_csv(r'../data/processed/target_values/test_target_values.csv', header=None)

### Build and Train Models On Training Set 

adb_reg, train_adb_score = train_models.train_adb(train_preprocessed,train_target)

etr_reg, train_etr_score = train_models.train_etr(train_preprocessed,train_target)

xgb_reg, train_xgb_score = train_models.train_xgb(train_preprocessed,train_target)

# save the models to disk
adb_filename = r'../models/adaboost_model.sav'
etr_filename = r'../models/extratrees_model.sav'
xgb_filename = r'../models/xgboost_model.sav'

pickle.dump(adb_reg, open(adb_filename, 'wb'))
pickle.dump(etr_reg, open(etr_filename, 'wb'))
pickle.dump(xgb_reg, open(xgb_filename, 'wb'))

### Stack and Blend Models
Stack the above 3 models and blend them using a Random Forest Regressor. Then train the blender on the validation set

blender_model, blender_score = train_models.stack_models(val_preprocessed, val_target, xgb_reg, adb_reg, etr_reg)

### Evaluate The Best Model On Test Set

# Create a table of all the models with their mean cross-validated scores
scoreboard = pd.DataFrame({'Model_Name': ['XGBoost Regressor', 'AdaBoost Regressor', 'Extra Trees Regressor', 'RF Blender'],
                           'Model': [xgb_reg, adb_reg, etr_reg, blender_model],
                           'Score': [train_xgb_score, train_adb_score, train_etr_score, blender_score]})

winner_model_name = scoreboard.loc[scoreboard[['Score']].idxmin()]['Model_Name']
winner_model = scoreboard.loc[scoreboard[['Score']].idxmin()]['Model']

print (scoreboard)
print('{} model performs the best on Validation Set'.format(winner_model_name[0]))

# Check the Root of Winner Model on Test set
final_score = train_models.evaluate_on_test(winner_model[0], test_preprocessed, test_target)

print('The final Root Mean Squared Error (RMSE score) of {} on unseen data is {}'.format(winner_model_name[0], np.round(final_score, 2)))

### Conclusion
Here, I compared three ensemble algorithms, viz. XGBoost, AdaBoost and Extremely Randomized Trees. I also stacked these models and blended them using a Random Forest Regressor. Though XGBoost and Extra Trees Regressors have comparable performances, XGBoost performed better than all others.

Also, these models have been trained on a subset of the original data (size of the dataset is 60k rows, considering the processing/ memory capabilities of my machine). If we train the models on a larger size of original data or the entire original data then the performance of these models would certainly improve.

I will update the models and their scores once I get train it on the entire dataset.

## Thanks for reading!
